move conflict function from git to render
