+++ src/js/gitnative.js
@@ -2,7 +2,6 @@
 
 var path = require('path'),
     utils = require('../js/lib/utils'),
-    fs = require('fs'),
     cmdr = require('../js/lib/cmdr');
 
 // NOTE: could use moment module for this
@@ -173,9 +172,6 @@ function GitManager() {
                     }
                 })
                 .then(function () {
-                    if (conflicts) {
-                        mythis.consolidateConflicts(localPath, conflicts);
-                    }
                     return utils.fs.outputFile(localManifestPath, toJSON(mergedManifest));
                 })
                 .then(function () {
@@ -189,37 +185,6 @@ function GitManager() {
                     return conflicts;
                 });
 
-        },
-
-        consolidateConflicts: function (targetPath, conflicts) {
-            var conflicttest = new RegExp(/([^<=>]*)(<{7} HEAD\n)([^<=>]*)(={7}\n)([^<=>]*)(>{7} \w{40}\n?)([^]*)/);
-
-            conflicts.forEach(function (conflict) {
-                var split = conflict.split("-");
-                var filePath = path.join(targetPath, split[0], split[1] + ".txt");
-                var contents = fs.readFileSync(filePath, "utf8");
-                var head = "";
-                var middle = "";
-                var tail = "";
-                var first = "";
-                var second = "";
-
-                while (conflicttest.test(contents)) {
-                    var pieces = conflicttest.exec(contents);
-
-                    head = pieces[2];
-                    middle = pieces[4];
-                    tail = pieces[6];
-                    first += pieces[1] + pieces[3];
-                    second += pieces[1] + pieces[5];
-                    contents = pieces[7];
-                }
-
-                first += contents;
-                second += contents;
-                var newcontent = head + first + middle + second + tail;
-                fs.writeFileSync(filePath, newcontent);
-            });
         },
 
         push: function (user, dir, repo, opts) {
