+++ src/js/render.js
@@ -113,6 +113,37 @@ function Renderer() {
             }
         },
 
+        consolidateConflicts: function (targetPath, conflicts) {
+            var conflicttest = new RegExp(/([^<=>]*)(<{7} HEAD\n)([^<=>]*)(={7}\n)([^<=>]*)(>{7} \w{40}\n?)([^]*)/);
+
+            conflicts.forEach(function (conflict) {
+                var split = conflict.split("-");
+                var filePath = path.join(targetPath, split[0], split[1] + ".txt");
+                var contents = fs.readFileSync(filePath, "utf8");
+                var head = "";
+                var middle = "";
+                var tail = "";
+                var first = "";
+                var second = "";
+
+                while (conflicttest.test(contents)) {
+                    var pieces = conflicttest.exec(contents);
+
+                    head = pieces[2];
+                    middle = pieces[4];
+                    tail = pieces[6];
+                    first += pieces[1] + pieces[3];
+                    second += pieces[1] + pieces[5];
+                    contents = pieces[7];
+                }
+
+                first += contents;
+                second += contents;
+                var newcontent = head + first + middle + second + tail;
+                fs.writeFileSync(filePath, newcontent);
+            });
+        },
+
         replaceEscapes: function (text) {
             text = text.replace(/\\/g, "\\\\").replace(/\[/g, "\\[").replace(/\]/g, "\\]").replace(/\^/g, "\\^").replace(/\$/g, "\\$");
             text = text.replace(/\(/g, "\\(").replace(/\)/g, "\\)").replace(/\?/g, "\\?").replace(/\./g, "\\.").replace(/\//g, "\\/");
